(original rev 3295) merge dev into staging
Merged revisions 3208-3212,3215-3217,3219-3220,3224-3225,3229-3248,3252-3264,3267-3284,3286-3292 via svnmerge from
file:///gaba/svn/netgpu/omni/dev

........
  r3208 |  | 2010-10-14 21:16:33 -0400 (Thu, 14 Oct 2010) | 1 line

  refactor core segment data itno struct
........
  r3209 |  | 2010-10-14 21:20:39 -0400 (Thu, 14 Oct 2010) | 1 line

  OmId --> OmID
........
  r3210 |  | 2010-10-14 21:59:11 -0400 (Thu, 14 Oct 2010) | 1 line

  checkpoint segment work
........
  r3211 |  | 2010-10-14 22:14:47 -0400 (Thu, 14 Oct 2010) | 1 line

  checkpoint segment memmap
........
  r3212 |  | 2010-10-14 22:17:31 -0400 (Thu, 14 Oct 2010) | 1 line

  checkpoint segment memmap
........
  r3215 |  | 2010-10-14 23:13:31 -0400 (Thu, 14 Oct 2010) | 1 line

  with memmap, no longer need to keep track of dirty segments
........
  r3216 |  | 2010-10-14 23:13:50 -0400 (Thu, 14 Oct 2010) | 1 line

  whitespace
........
  r3217 |  | 2010-10-15 00:20:59 -0400 (Fri, 15 Oct 2010) | 1 line

  checkpoint mem map conversion from hdf5
........
  r3219 |  | 2010-10-15 00:30:34 -0400 (Fri, 15 Oct 2010) | 1 line

  dont mix text w/ datastream
........
  r3220 |  | 2010-10-15 00:35:59 -0400 (Fri, 15 Oct 2010) | 1 line

  time load of segment pages
........
  r3224 |  | 2010-10-15 00:47:40 -0400 (Fri, 15 Oct 2010) | 1 line

  bug fix for versions
........
  r3225 |  | 2010-10-15 00:49:31 -0400 (Fri, 15 Oct 2010) | 1 line

  fiddle w/ versions
........
  r3229 |  | 2010-10-15 01:02:32 -0400 (Fri, 15 Oct 2010) | 1 line

  tiny refactor; dont need double pass when converting from old segments
........
  r3230 |  | 2010-10-15 01:15:31 -0400 (Fri, 15 Oct 2010) | 1 line

  fix path issue when upgrading old file
........
  r3231 |  | 2010-10-15 01:38:11 -0400 (Fri, 15 Oct 2010) | 1 line

  first unit test for paths
........
  r3232 |  | 2010-10-15 01:38:15 -0400 (Fri, 15 Oct 2010) | 1 line

  first unit test for paths
........
  r3233 |  | 2010-10-15 02:06:33 -0400 (Fri, 15 Oct 2010) | 1 line

  small cleanup of shared_ptr wrappers
........
  r3234 | purcaro | 2010-10-15 14:15:49 -0400 (Fri, 15 Oct 2010) | 1 line

  small refactor
........
  r3235 | purcaro | 2010-10-15 14:18:11 -0400 (Fri, 15 Oct 2010) | 1 line

  comment
........
  r3236 | purcaro | 2010-10-15 14:19:53 -0400 (Fri, 15 Oct 2010) | 1 line

  tweak
........
  r3237 | purcaro | 2010-10-15 14:27:45 -0400 (Fri, 15 Oct 2010) | 1 line

  tweak exception handling
........
  r3238 | purcaro | 2010-10-15 14:31:26 -0400 (Fri, 15 Oct 2010) | 1 line

  whitespace
........
  r3239 | purcaro | 2010-10-15 14:47:25 -0400 (Fri, 15 Oct 2010) | 1 line

  fix bug w/ getting segment that doesnt exist
........
  r3240 | purcaro | 2010-10-15 14:58:38 -0400 (Fri, 15 Oct 2010) | 1 line

  tweak
........
  r3241 | purcaro | 2010-10-15 15:03:16 -0400 (Fri, 15 Oct 2010) | 1 line

  remove unneeded function
........
  r3242 | purcaro | 2010-10-15 15:07:30 -0400 (Fri, 15 Oct 2010) | 1 line

  cleanup
........
  r3243 | purcaro | 2010-10-15 15:14:56 -0400 (Fri, 15 Oct 2010) | 1 line

  make more obvious
........
  r3244 | purcaro | 2010-10-15 15:15:43 -0400 (Fri, 15 Oct 2010) | 1 line

  tweak
........
  r3245 | purcaro | 2010-10-15 15:26:17 -0400 (Fri, 15 Oct 2010) | 1 line

  comment
........
  r3246 |  | 2010-10-15 17:00:12 -0400 (Fri, 15 Oct 2010) | 1 line

  make one segments folder per segmentation on disk
........
  r3247 |  | 2010-10-15 17:16:11 -0400 (Fri, 15 Oct 2010) | 1 line

  split off paging store metadata
........
  r3248 |  | 2010-10-15 17:30:00 -0400 (Fri, 15 Oct 2010) | 1 line

  nuke dead upgrader
........
  r3252 |  | 2010-10-15 17:44:04 -0400 (Fri, 15 Oct 2010) | 1 line

  check that we are at end of stream
........
  r3253 |  | 2010-10-16 01:11:24 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint mst
........
  r3254 |  | 2010-10-16 01:16:04 -0400 (Sat, 16 Oct 2010) | 1 line

  tweak for user threshold
........
  r3255 |  | 2010-10-16 01:18:04 -0400 (Sat, 16 Oct 2010) | 1 line

  keep graph consistent
........
  r3256 |  | 2010-10-16 02:08:37 -0400 (Sat, 16 Oct 2010) | 1 line

  unmap when destroyfing
........
  r3257 |  | 2010-10-16 02:59:27 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint reworking downsampler
........
  r3258 |  | 2010-10-16 22:31:56 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint downsampler
........
  r3259 |  | 2010-10-16 22:34:03 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint downsampler
........
  r3260 |  | 2010-10-16 22:37:43 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint downsampler
........
  r3261 |  | 2010-10-16 22:49:53 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint downsampler
........
  r3262 |  | 2010-10-16 23:40:53 -0400 (Sat, 16 Oct 2010) | 1 line

  checkpoint downsampler
........
  r3263 |  | 2010-10-16 23:49:56 -0400 (Sat, 16 Oct 2010) | 1 line

  cache chunk location
........
  r3264 |  | 2010-10-17 00:18:55 -0400 (Sun, 17 Oct 2010) | 1 line

  bug fix
........
  r3267 |  | 2010-10-18 13:55:41 -0400 (Mon, 18 Oct 2010) | 1 line

  whitespace
........
  r3268 |  | 2010-10-18 13:56:38 -0400 (Mon, 18 Oct 2010) | 1 line

  add skeleton
........
  r3269 |  | 2010-10-18 14:51:19 -0400 (Mon, 18 Oct 2010) | 1 line

  bug fix
........
  r3270 |  | 2010-10-18 14:54:02 -0400 (Mon, 18 Oct 2010) | 1 line

  use uint64_T
........
  r3271 |  | 2010-10-18 15:44:47 -0400 (Mon, 18 Oct 2010) | 1 line

  switch from float to double
........
  r3272 |  | 2010-10-18 17:12:55 -0400 (Mon, 18 Oct 2010) | 1 line

  first attempt at testing segment graph
........
  r3273 |  | 2010-10-18 18:33:42 -0400 (Mon, 18 Oct 2010) | 3 lines

  tests for segments and refactor
........
  r3274 |  | 2010-10-18 23:47:12 -0400 (Mon, 18 Oct 2010) | 2 lines

  shortcut 0 segment ID for getsegment; test for valdiity of segment ID; zero out memory for fake mem map
........
  r3275 |  | 2010-10-18 23:50:07 -0400 (Mon, 18 Oct 2010) | 1 line

  superfluous call
........
  r3276 |  | 2010-10-19 00:09:59 -0400 (Tue, 19 Oct 2010) | 1 line

  create test project file
........
  r3277 |  | 2010-10-19 00:12:48 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3278 |  | 2010-10-19 00:19:42 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3279 |  | 2010-10-19 00:23:49 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3280 |  | 2010-10-19 00:29:30 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3281 |  | 2010-10-19 00:29:43 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3282 |  | 2010-10-19 00:30:47 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3283 |  | 2010-10-19 00:34:47 -0400 (Tue, 19 Oct 2010) | 1 line

  small cleanup of tests
........
  r3284 |  | 2010-10-19 00:48:52 -0400 (Tue, 19 Oct 2010) | 1 line

  use variable for max segment num
........
  r3286 |  | 2010-10-19 02:37:10 -0400 (Tue, 19 Oct 2010) | 1 line

  more float to double conversion in segments; nuke dead code
........
  r3287 |  | 2010-10-19 02:41:36 -0400 (Tue, 19 Oct 2010) | 1 line

  typo
........
  r3288 |  | 2010-10-19 03:06:35 -0400 (Tue, 19 Oct 2010) | 1 line

  playing w/ segment colorizer; untested
........
  r3289 |  | 2010-10-19 03:47:01 -0400 (Tue, 19 Oct 2010) | 1 line

  play with makeSelectedColor
........
  r3290 |  | 2010-10-19 11:30:17 -0400 (Tue, 19 Oct 2010) | 1 line

  forgot file
........
  r3291 |  | 2010-10-19 15:46:28 -0400 (Tue, 19 Oct 2010) | 1 line

  sync w/ aleks
........
  r3292 |  | 2010-10-19 16:34:01 -0400 (Tue, 19 Oct 2010) | 1 line

  switch mst and segment data lists from mem-map to regular file access
........


git-svn-id: file:///tmp/current/omni/branches/staging@3048 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/segment/omSegmentEdge.cpp
@@ -16,7 +16,7 @@ OmSegmentEdge::OmSegmentEdge( OmSegment * c )
 {
 }
 
-OmSegmentEdge::OmSegmentEdge( const OmSegID p, const OmSegID c, const float t )
+OmSegmentEdge::OmSegmentEdge( const OmSegID p, const OmSegID c, const double t )
 	: parentID(p)
 	, childID(c)
 	, threshold(t)
@@ -24,7 +24,7 @@ OmSegmentEdge::OmSegmentEdge( const OmSegID p, const OmSegID c, const float t )
 {
 }
 
-OmSegmentEdge::OmSegmentEdge( OmSegment * p, OmSegment * c, const float t )
+OmSegmentEdge::OmSegmentEdge( OmSegment * p, OmSegment * c, const double t )
 	: parentID(p->value())
 	, childID(c->value())
 	, threshold(t)
